pyproject.toml
[project]
name = "this-little-wiggy"
version = "0.1.0"
description = "Claude Code plugin that formats complex tasks for ralph-wiggum autonomous loop"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "severity1" }
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pyyaml>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
